refactor(sqlite): implement QueryEngine on SqliteEngine
SqliteEngine used to return JSON blobs with no path to typed models.
Now it deserializes rows via FromRow + SqliteRowRef. The legacy JSON
API moves to prax_sqlite::raw (SqliteRawEngine / SqliteJsonRow) for
backward compatibility.

Changes:

- **engine.rs**: Full rewrite implementing prax_query::QueryEngine.
  Uses SqliteRowRef (from Task 6) to materialize rows, then
  T::from_row() to deserialize into typed models. Returns typed T
  instead of JSON.

- **raw.rs**: New module preserving the old JSON-returning API
  (SqliteRawEngine, SqliteJsonRow). Renamed from SqliteEngine /
  SqliteQueryResult to avoid conflicts. All legacy method signatures
  unchanged.

- **lib.rs**: Added raw module, updated re-exports. New code uses
  SqliteEngine, legacy code uses SqliteRawEngine.

- **types.rs**: filter_value_to_sqlite already handles all seven
  FilterValue variants (Null, Bool, Int, Float, String, Json, List).
  No changes needed.

- **tests/engine_crud.rs**: New integration test verifying
  QueryEngine::query_many() deserializes rows into typed Item structs.
  Uses shared in-memory database to avoid SQLite per-connection
  isolation issues.

- **benches/sqlite_operations.rs**: Updated to use SqliteRawEngine
  instead of SqliteEngine (benchmark tests legacy JSON API).

Breaking change: SqliteEngine no longer returns JSON. Use
SqliteRawEngine for the old API or migrate to typed models.
