+++ prax-sqlite/src/lib.rs
@@ -31,13 +31,15 @@ pub mod connection;
 pub mod engine;
 pub mod error;
 pub mod pool;
+pub mod raw;
 pub mod row;
 pub mod row_ref;
 pub mod types;
 
 pub use config::{DatabasePath, JournalMode, SqliteConfig, SynchronousMode};
 pub use connection::SqliteConnection;
-pub use engine::{SqliteEngine, SqliteQueryResult};
+pub use engine::SqliteEngine;
 pub use error::{SqliteError, SqliteResult};
 pub use pool::{PoolConfig, SqlitePool, SqlitePoolBuilder};
+pub use raw::{SqliteJsonRow, SqliteRawEngine};
 pub use row::FromSqliteRow;
